pyproject.toml
[project]
name = "photobooth"
version = "0.1.0"
description = "we are going to take pibooth and upgrade the hell out of it. and make it into a modern product that runs great on new raspberry pis. i have a photo booth we can use to test! "
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "tomli_w>=1.0.0",
    "pluggy>=1.3.0",
    "Pillow>=10.4.0",
    "qrcode[pil]>=7.4",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
camera = ["opencv-python>=4.8.0"]
gpio = ["gpiozero>=2.0"]
printer = ["pycups>=2.0.0"]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"